+++ scripts/test/Selenium/Agent/Admin/AdminNotificationEvent.t
@@ -119,6 +119,31 @@ $Selenium->RunTest(
         $Selenium->find_element( '#MIMEBase_Subject', 'css' )->send_keys($NotifEventText);
         $Selenium->find_element( '#en_Subject',       'css' )->send_keys($NotifEventText);
 
+        # Check 'Additional recipient' length validation from Additional recipient email addresses (see bug#13936).
+        my $FieldValue = "a" x 201;
+
+        # Check TransportEmail checkbox if it is not checked.
+        my $TransportEmailCheck = $Selenium->execute_script("return \$('#TransportEmail').prop('checked');");
+        if ( !$TransportEmailCheck ) {
+            $Selenium->execute_script("\$('#TransportEmail').prop('checked', true);");
+            $Selenium->WaitFor( JavaScript => "return \$('#TransportEmail').prop('checked') === true;" );
+        }
+        $Selenium->find_element( "#RecipientEmail", 'css' )->send_keys($FieldValue);
+        $Selenium->find_element( "#Submit",         'css' )->click();
+        $Selenium->WaitFor( JavaScript => "return \$('#RecipientEmail.Error').length;" );
+
+        $Self->True(
+            $Selenium->execute_script("return \$('#RecipientEmail.Error').length;"),
+            "Validation for 'Additional recipient' field is correct",
+        );
+        $Selenium->find_element( "#RecipientEmail", 'css' )->clear();
+
+        # Set back TransportEmail checkbox if it was not checked.
+        if ( !$TransportEmailCheck ) {
+            $Selenium->execute_script("\$('#TransportEmail').prop('checked', false);");
+            $Selenium->WaitFor( JavaScript => "return \$('#TransportEmail').prop('checked') === false;" );
+        }
+
         # Insert long string into text area using jQuery, since send_keys() takes too long.
         $Selenium->execute_script(
             "\$('#en_Body').val('$TooLongString').trigger('change');"